requirements.txt

python-dotenv==0.8.2
spotipy==2.4.4
